backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
pandas>=2.2.1
pydantic>=2.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
MetaTrader5>=5.0.0
yfinance>=0.2.37
